pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "escai"
version = "0.2.0"
description = "Epistemic State and Causal Analysis Intelligence Framework for monitoring autonomous agent cognition"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "ESCAI Framework Team"}
]
keywords = ["ai", "agents", "monitoring", "observability", "cognition"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core processing (minimal for basic functionality)
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    
    # CLI dependencies
    "click>=8.0.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
    
    # Utilities
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
]

test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.1.0",
    "pytest-mock>=3.10.0",
    "pytest-benchmark>=4.0.0",
    "coverage>=7.0.0",
    "aiohttp>=3.8.0",
    "psutil>=5.9.0",
]

# Extended dependencies for full functionality
full = [
    # Core processing
    "scikit-learn>=1.1.0",
    "xgboost>=1.6.0",
    "networkx>=2.8.0",
    
    # NLP & AI
    "transformers>=4.20.0",
    "nltk>=3.7",
    "sentence-transformers>=2.2.0",
    "torch>=1.12.0",
    
    # API & Communication
    "fastapi>=0.95.0",
    "websockets>=11.0",
    "pydantic>=1.10.0",
    "uvicorn>=0.20.0",
    
    # Storage & Databases
    "sqlalchemy>=1.4.0",
    "redis>=4.3.0",
    "pymongo>=4.2.0",
    "psycopg2-binary>=2.9.0",
    "influxdb-client>=1.30.0",
    "neo4j>=5.0.0",
    
    # Visualization
    "plotly>=5.10.0",
    "streamlit>=1.20.0",
    "matplotlib>=3.5.0",
    
    # Utilities
    "python-multipart>=0.0.5",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.scripts]
escai = "escai_framework.cli.main:main"

[project.urls]
Homepage = "https://github.com/Sonlux/ESCAI"
Documentation = "https://sonlux.github.io/ESCAI"
Repository = "https://github.com/Sonlux/ESCAI"
"Bug Tracker" = "https://github.com/Sonlux/ESCAI/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["escai_framework*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --strict-config"
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "performance: Performance tests",
    "load: Load tests",
    "accuracy: Accuracy validation tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
    "requires_db: Tests that require database connection",
    "requires_network: Tests that require network access"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]